interim changes

+++ packages/auth/src/route/createClientRoute.ts
@@ -21,13 +21,16 @@ export const createClientRoute: () => express.Router = () => {
     const user = req.user as User;
     const user_id = user.id;
 
-    if (!application_name || !client_secret)
-      return res.status(httpStatus.BAD_REQUEST).send({ error: 'application_name, client_secret is required' });
+    if (!application_name || !client_secret) {
+      logger.warn(`cannot create client: missing params - application_name, client_secret`);
+      return res.status(httpStatus.BAD_REQUEST).send({ error: 'missing params - application_name, client_secret' });
+    }
 
     const client = Client.create({ application_name, client_secret, redirect_uris, grants, user_id, is_system_app });
 
     try {
       await Client.insert(client);
+      logger.info(`account ${user_id} register new client ${client.id}`);
       return res.status(httpStatus.OK).send({ ok: true, application_name, id: client.id });
     } catch (e) {
       logger.error(util.format('fail to insert client %s, %j', client.id, e));
@@ -43,6 +46,7 @@ export const createClientRoute: () => express.Router = () => {
     if (application_name) {
       try {
         const client = await Client.findOne({ application_name, user_id: user.id });
+        logger.info(`account ${user.id} retrieves client record ${client.id}`);
         return res.status(httpStatus.OK).send(omit(client, 'client_secret'));
       } catch (e) {
         logger.error(util.format('fail to retrieve client, %j', e));
@@ -56,6 +60,7 @@ export const createClientRoute: () => express.Router = () => {
           skip: 0,
           take: 10
         });
+        logger.info(`account ${user.id} retrieves ${clients.length} client records`);
         return res.status(httpStatus.OK).send(clients);
       } catch (e) {
         logger.error(util.format('fail to retrieve list of clients, %j', e));
@@ -70,6 +75,7 @@ export const createClientRoute: () => express.Router = () => {
 
     try {
       const client = await Client.findOne({ where: { id: client_id, user_id: user.id } });
+      logger.info(`account ${user.id} retrieves client ${client_id}`);
       return res.status(httpStatus.OK).send(client);
     } catch (e) {
       logger.error(util.format('fail to retrieve client %s, %j', client_id, e));
@@ -79,17 +85,30 @@ export const createClientRoute: () => express.Router = () => {
 
   router.put('/:client_id', passport.authenticate('bearer', { session: false }), async (req, res) => {
     const client_id = req.params.client_id;
+    const user = req.user as User;
+
+    if (!req.body?.application_name && !req.body?.redirect_uris && !req.body?.grants && !req.body?.client_secret) {
+      logger.warn(
+        `cannot update client ${client_id}: missing params: application_name, client_secret, redirect_uris, or grants`
+      );
+      return res
+        .status(httpStatus.BAD_REQUEST)
+        .send({ error: 'missing params: application_name, client_secret, redirect_uris, or grants' });
+    }
 
     let client: Client;
 
     try {
-      client = await Client.findOne({ where: { client_id } });
+      client = await Client.findOne({ where: { id: client_id, user_id: user.id } });
     } catch (e) {
       logger.error(util.format('fail to find client %s, %j', client_id, e));
       return res.status(httpStatus.BAD_REQUEST).send({ error: 'fail to find client' });
     }
 
-    if (!client) return res.status(httpStatus.BAD_REQUEST).send({ error: 'fail to find client' });
+    if (!client) {
+      logger.warn(`cannot update ${client_id}: fail to find client`);
+      return res.status(httpStatus.BAD_REQUEST).send({ error: 'fail to find client' });
+    }
 
     try {
       const payload = {
@@ -99,6 +118,7 @@ export const createClientRoute: () => express.Router = () => {
         grants: req.body?.grants || client.grants
       };
       await Client.update(client_id, payload);
+      logger.info(`account ${user.id} updates client ${client_id}`);
       return res.status(httpStatus.OK).send({ ok: true, ...payload });
     } catch (e) {
       logger.error(util.format('fail to update client %s, %j', client_id, e));
@@ -110,8 +130,23 @@ export const createClientRoute: () => express.Router = () => {
     const client_id = req.params.client_id;
     const user = req.user as User;
 
+    let client: Client;
+
+    try {
+      client = await Client.findOne({ where: { id: client_id, user_id: user.id } });
+    } catch (e) {
+      logger.error(util.format('fail to find client %s, %j', client_id, e));
+      return res.status(httpStatus.BAD_REQUEST).send({ error: 'fail to find client' });
+    }
+
+    if (!client) {
+      logger.warn(`cannot delete ${client_id}: fail to find client`);
+      return res.status(httpStatus.BAD_REQUEST).send({ error: 'fail to find client' });
+    }
+
     try {
       await Client.delete(client_id);
+      logger.info(`account ${user.id} deletes client ${client_id}`);
       return res.status(httpStatus.OK).send({ ok: true });
     } catch (e) {
       logger.error(util.format('fail to delete client %s, %j', client_id, e));
